backend/requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0

# ML and Image Processing dependencies
numpy==1.23.5
scipy==1.10.1
tensorflow==2.12.0
keras==2.12.0
pillow==10.0.1

# LLM dependencies
torch==2.0.1
transformers==4.33.3
accelerate==0.23.0

# Testing and Development
pytest==7.4.2
httpx==0.25.0
gunicorn==21.2.0 